+++ packages/nextjs/components/scaffold-eth/index.tsx
@@ -6,3 +6,4 @@ export * from "./Faucet";
 export * from "./FaucetButton";
 export * from "./Input";
 export * from "./RainbowKitCustomConnectButton";
+export * from "./DonateButton";
